Add DB API functions and major refactor
The DB interface is mostly done, but there's some smelly with something in common:
it tries to overcome sqldelight limits for this project's purpose. It turns out
sqldelight is not that great if your plan is to swap database implementations
seamlessly: I don't want to expose its dataclasses to the core of the application,
but each query will build its own one even if they share the exact same signature,
which means I have to be commited to write a lot of boilerplate code to convert
those into my own dataclasses.

On the other hand I cannot find a flexible way to do updates on a variable number
of columns.

This looks like a great library for Android where you'll probably ever use sqlite,
but not for this.