Split responsabilities among exceptions
The new `UnsupportedDownloadException` is meant to be used in case
of errors found during the image processing, except the underlying
IO operations to reach the URL (hence if a stream to the URL can't be
established, it will thrown an `IOException`.)

This class also provides error codes in the form of an enum class,
`DownloadErrorType`. As of right now, to codes are listed: one used
in the case of an image being too big for Sheska to process, and
another one in case the image file isn't supported at all by
`ImageIO`.