Enhance status message options
This is a quick change without touching `MangaExtension`. It's
often preferable to send an ephemeral response instead of a public
one, and it's often preferable to have ephemeral commands instead of
public ones, so that's what these new functions are for. Nonetheless,
it'd be much better if there were only two functions instead of six,
to be called like `respondStatus(Status.ERROR, ...)` or similar.