Add converter class for generic lists
This takes a string like `n1, n2, n3, n4` where n can be
converted to a type `T` with a function `(String) -> T?`.

A returning `null` value from the converting function means
at least one conversions for any n was unsuccesful,
invalidating the entire argument.

The function that converts each of the substrings should
take care of exceptions by itself - the converter class
will not catch them, and it only expects `T?` values from it.

The original input string is splitted into substrings using
a comma `','` as separator. This behaviour can't be changed.